Split Factory into smaller semantic enums
