+++ Fresco/Fresco.swift
@@ -19,7 +19,7 @@ public struct Fresco: ParsableCommand {
     public init() {}
 
     public func run() throws {
-      let target = try Factory.create(from: target)
+      let target = try TargetFactory.create(from: target)
       try IconManager.reset(for: target)
     }
   }
@@ -44,8 +44,8 @@ public struct Fresco: ParsableCommand {
     public init() {}
 
     public func run() throws {
-      let resource = try Factory.create(from: resource)
-      let target = try Factory.create(from: target)
+      let resource = try ResourceFactory.create(from: resource)
+      let target = try TargetFactory.create(from: target)
       try IconManager.set(resource, for: target)
     }
   }
